Move encryption_helper to utils/ dir.
