+++ gslib/commands/compose.py
@@ -22,12 +22,12 @@ from gslib.bucket_listing_ref import BucketListingObject
 from gslib.command import Command
 from gslib.command_argument import CommandArgument
 from gslib.cs_api_map import ApiSelector
-from gslib.encryption_helper import GetEncryptionKeyWrapper
 from gslib.exception import CommandException
 from gslib.storage_url import ContainsWildcard
 from gslib.storage_url import StorageUrlFromString
 from gslib.third_party.storage_apitools import storage_v1_messages as apitools_messages
 from gslib.translation_helper import PreconditionsFromHeaders
+from gslib.utils.encryption_helper import GetEncryptionKeyWrapper
 
 MAX_COMPONENT_COUNT = 1024
 MAX_COMPOSE_ARITY = 32
