+++ gslib/tests/test_rewrite.py
@@ -24,7 +24,6 @@ from boto.storage_uri import BucketStorageUri
 
 from gslib.cs_api_map import ApiSelector
 from gslib.discard_messages_queue import DiscardMessagesQueue
-from gslib.encryption_helper import CryptoKeyWrapperFromKey
 from gslib.gcs_json_api import GcsJsonApi
 from gslib.project_id import PopulateProjectId
 from gslib.tests.rewrite_helper import EnsureRewriteRestartCallbackHandler
@@ -43,6 +42,7 @@ from gslib.tests.util import TEST_ENCRYPTION_KEY4
 from gslib.tests.util import unittest
 from gslib.tracker_file import DeleteTrackerFile
 from gslib.tracker_file import GetRewriteTrackerFilePath
+from gslib.utils.encryption_helper import CryptoKeyWrapperFromKey
 from gslib.utils.unit_util import ONE_MIB
