+++ gslib/tests/util.py
@@ -31,10 +31,10 @@ import boto
 import crcmod
 from gslib.cloud_api import ResumableDownloadException
 from gslib.cloud_api import ResumableUploadException
-from gslib.encryption_helper import Base64Sha256FromBase64EncryptionKey
 from gslib.lazy_wrapper import LazyWrapper
 import gslib.tests as gslib_tests
 from gslib.utils.boto_util import UsingCrcmodExtension
+from gslib.utils.encryption_helper import Base64Sha256FromBase64EncryptionKey
 from gslib.utils.posix_util import GetDefaultMode
 from gslib.utils.system_util import IS_WINDOWS
 from gslib.utils.unit_util import MakeHumanReadable
