+++ gslib/utils/encryption_helper.py
